refactor(desktop): satisfy inbox lint budgets
Problem:
The inbox fix exceeded repository file-size and function-size budgets in CI.

Root cause:
New reconciliation tests and cancellation logic were added to legacy files that were already at or near their ratcheted limits.

Fix:
Extract cancellation, inbox event, steer loader, and regression-test ownership into focused files. Keep only a two-line bridge contract ratchet for the additive Wails method.

Verification:
- go run ./tools/repolint
- go test ./internal/agent ./internal/control ./internal/sessioninbox
- go test -race ./internal/sessioninbox
- cd desktop && go test ./...
- frontend type checks, hook lint, focused tests, and production build
