+++ desktop/frontend/src/components/Composer.tsx
@@ -28,7 +28,8 @@ import {
   type StructuredInvocationSubmit,
 } from "../lib/invocationDisplay";
 import { formatTokens } from "../lib/format";
-import type { CancelOutcome, ControllerLiveStore } from "../lib/useController";
+import type { CancelOutcome } from "../lib/inboxCancel";
+import type { ControllerLiveStore } from "../lib/useController";
 import { clearLayoutSize, loadOptionalLayoutSize, saveLayoutSize } from "../lib/layoutPreferences";
 import { createRafResizeUpdater } from "../lib/resizeDrag";
 import { observeComposerMenuViewport } from "../lib/composerMenuViewport";
