+++ tools/repolint/baseline.json
@@ -151,7 +151,7 @@
       "file-size": 415
     },
     "desktop/frontend/src/lib/bridge.ts": {
-      "file-size": 4780
+      "file-size": 4782
     },
     "desktop/frontend/src/lib/crash.ts": {
       "file-size": 225
